Initial design, starting with a `Floor` calling for an elevator.
Going with an event-driven design, where a `Floor` publishes
an `ElevatorRequestUpEvent` to request an elevator going UP and an
`ElevatorRequestUpEvent` to request an elevator going DOWN.

This initial commit also contains the groundwork and scaffolding
for a running application.  You can either build and run in 1 step:

```
$ ./gradlew bootRun
```

Or you can take the "production" approach and build, then run
as an executable jar.

Build:
```
$ ./gradlew build
```

Then run:
```
$ java -jar build/libs/elevator-demo-0.0.1.BUILD-SNAPSHOT.jar
```

In both cases, the build will also implicitly build and run the tests.