Added `Elevator`s that can move to a particular floor.
When an elevator moves , it will produce the following events:

+ ElevatorMovingEvent
+ ElevatorStatusUpdateEvent

If it passes by a floor (without stopping):
+ `ElevatorFloorPassEvent`

If it stops at a floor:
+ `ElevatorFloorStopEvent`
+ `ElevatorDoorOpenEvent`
+ `ElevatorDoorCloseEvent`
+ `ElevatorDiagnosticUpdateEvent` // at every floor stop

in that order.  The movement simulation is implemented as a
reactive `Flux` that emits signals as "floors visited".  We can
also inject a configurable delay between each emission, which
emulates the time it takes for the elevator to move between floors.
For testing, we collapse this delay to zero.